refactor: Console 추상화
